Updated global exception handler to handle binding results for DTOs. Restructured user entity package.
