+++ Rent/Rent-Data/src/main/java/dataAccess/entities/user/User.java
@@ -1,4 +1,4 @@
-package dataAccess.entities;
+package dataAccess.entities.user;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.hibernate.annotations.Type;
@@ -91,6 +91,7 @@ public class User implements java.io.Serializable {
     }
 
     @Basic
+    @JsonIgnore
     @Column(name = "lastName")
     public String getLastName() {
 
@@ -102,6 +103,7 @@ public class User implements java.io.Serializable {
     }
 
     @Basic
+    @JsonIgnore
     @Column(name = "firstName")
     public String getFirstName() {
