+++ Rent/Rent-Main/src/main/java/Controllers/GeneralController.java
@@ -1,9 +1,8 @@
 package Controllers;
 
 import dataAccess.dao.User.UserDAO;
-import dataAccess.entities.User;
+import dataAccess.entities.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
