+++ Rent/Rent-Main/src/main/java/Services/Login/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package Services.Login;
 
 import dataAccess.dao.User.UserDAO;
-import dataAccess.entities.User;
+import dataAccess.entities.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.security.core.GrantedAuthority;
