+++ Rent/Rent-Util/src/main/java/Utilities/Constants.java
@@ -6,4 +6,8 @@ package Utilities;
 public class Constants {
     // REGISTRATION TOKEN CONSTANTS
     public static final int TOKEN_TIME_DURATION = 60*24;
+
+
+    // CONTENT TYPES
+    public static final String CONTENT_TYPE_JSON = "content-type=application/json";
 }
